인증 인가 추가

+++ src/main/java/com/sparta/schedule/service/CommentService.java
@@ -4,7 +4,9 @@ import com.sparta.schedule.dto.CommentRequestDto;
 import com.sparta.schedule.dto.CommentResponseDto;
 import com.sparta.schedule.entity.Comment;
 import com.sparta.schedule.entity.Schedule;
+import com.sparta.schedule.entity.User;
 import com.sparta.schedule.repository.CommentRepository;
+import com.sparta.schedule.security.UserDetailsImpl;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -20,8 +22,10 @@ public class CommentService {
         this.commentRepository = commentRepository;
     }
 
-    public CommentResponseDto createComment(@RequestBody CommentRequestDto requestDto) {
-        Comment comment = new Comment(requestDto);
+    public CommentResponseDto createComment(CommentRequestDto requestDto, UserDetailsImpl userDetails) {
+        User user = userDetails.getUser();
+
+        Comment comment = new Comment(requestDto, user);
         if (comment.getComment() == null){
             throw  new IllegalArgumentException("내용을 입력해주세요.");
         }
@@ -34,22 +38,36 @@ public class CommentService {
 
     }
 
-    public Long updateComment(Long commentid, CommentRequestDto requestDto){
+    public CommentResponseDto updateComment(Long commentId, CommentRequestDto requestDto, UserDetailsImpl userDetails){
+        User user = userDetails.getUser();
 
-        Comment comment = findComment(commentid);
-        comment.update(commentid, requestDto);
-        return commentid;
+        Comment comment = findComment(commentId);
+        if(!comment.getUser().getId().equals(user.getId())){
+            throw  new IllegalArgumentException("작성자만 수정할 수 있습니다.");
+        } else {
+            comment.update(commentId, requestDto);
+        }
+
+
+        CommentResponseDto commentResponseDto = new CommentResponseDto(comment);
+        return commentResponseDto;
     }
 
-    public String deleteComment(Long commentid, CommentRequestDto requestDto) {
+    public String deleteComment(Long commentId, UserDetailsImpl userDetails) {
+        User user = userDetails.getUser();
 
-        Comment comment = findComment(commentid);
-        commentRepository.delete(comment);
-        return "삭제되었습니다.";
+        Comment comment = findComment(commentId);
+        if(!comment.getUser().getId().equals(user.getId())){
+            throw  new IllegalArgumentException("작성자만 삭제할 수 있습니다.");
+        } else {
+            commentRepository.delete(comment);
+            return "삭제되었습니다.";
+        }
     }
 
-    private Comment findComment(Long commentid){
-        return commentRepository.findById(commentid).orElseThrow(() ->
+
+    private Comment findComment(Long commentId){
+        return commentRepository.findById(commentId).orElseThrow(() ->
                 new IllegalArgumentException("댓글이 존재하지 않습니다."));
     }
 }
